Move test_generate_responses_to_subnet_calls to batch_delivery.rs
